기본 실행 확인

+++ src/main/kotlin/Value.kt
@@ -6,11 +6,11 @@ import kotlin.math.*
 import kotlin.random.Random
 
 class Value(
-    var data: Double,
+    var data: Float,
     private val _children: Set<Value> = emptySet(),
     var _op: String = ""
 ) {
-    var grad: Double = 0.0
+    var grad: Float = 0.0f
     var _backward: () -> Unit = {}
 
     // --- 연산자 오버로딩 최적화 ---
@@ -24,7 +24,7 @@ class Value(
         return out
     }
 
-    operator fun plus(other: Number): Value = this + Value(other.toDouble())
+    operator fun plus(other: Number): Value = this + Value(other.toFloat())
 
     operator fun times(other: Value): Value {
         val out = Value(this.data * other.data, setOf(this, other), "*")
@@ -35,27 +35,27 @@ class Value(
         return out
     }
 
-    operator fun times(other: Number): Value = this * Value(other.toDouble())
+    operator fun times(other: Number): Value = this * Value(other.toFloat())
 
-    operator fun unaryMinus(): Value = this * -1.0
+    operator fun unaryMinus(): Value = this * -1.0f
 
     operator fun minus(other: Value): Value = this + -other
 
-    operator fun minus(other: Number): Value = this - Value(other.toDouble())
+    operator fun minus(other: Number): Value = this - Value(other.toFloat())
 
     // pow(-1) 대신 직접 나누기 구현
     operator fun div(other: Value): Value {
         val out = Value(this.data / other.data, setOf(this, other), "/")
         out._backward = {
-            this.grad += (1.0 / other.data) * out.grad
+            this.grad += (1.0f / other.data) * out.grad
             other.grad += (-this.data / (other.data * other.data)) * out.grad
         }
         return out
     }
 
-    operator fun div(other: Number): Value = this / Value(other.toDouble())
+    operator fun div(other: Number): Value = this / Value(other.toFloat())
 
-    fun pow(other: Double): Value {
+    fun pow(other: Float): Value {
         val out = Value(this.data.pow(other), setOf(this), "**$other")
         out._backward = {
             this.grad += (other * this.data.pow(other - 1)) * out.grad
@@ -66,15 +66,15 @@ class Value(
     // --- 활성화 함수 (Utils에서 이동) ---
 
     fun relu(): Value {
-        val out = Value(if (this.data < 0) 0.0 else this.data, setOf(this), "ReLU")
+        val out = Value(if (this.data < 0) 0.0f else this.data, setOf(this), "ReLU")
         out._backward = {
-            this.grad += (if (out.data > 0) 1.0 else 0.0) * out.grad
+            this.grad += (if (out.data > 0) 1.0f else 0.0f) * out.grad
         }
         return out
     }
 
     fun exp(): Value {
-        val out = Value(exp(this.data), setOf(this), "exp")
+        val out = Value(exp(this.data.toDouble()).toFloat(), setOf(this), "exp")
         out._backward = {
             this.grad += out.data * out.grad
         }
@@ -83,7 +83,7 @@ class Value(
 
     fun sigmoid(): Value {
         // sigmoid(x) = 1 / (1 + exp(-x))
-        val out = 1.0 / (1.0 + (-this).exp())
+        val out = 1.0f / (1.0f + (-this).exp())
         // _backward는 체인룰에 의해 자동으로 처리됨
         return out
     }
@@ -91,16 +91,16 @@ class Value(
     fun gelu(): Value {
         // GELU(x) = 0.5 * x * (1 + tanh(sqrt(2/π) * (x + 0.044715 * x^3)))
         val x = this
-        val c1 = Value(sqrt(2.0 / PI))
-        val c2 = Value(0.044715)
+        val c1 = Value(sqrt(2.0 / PI).toFloat())
+        val c2 = Value(0.044715f)
 
-        val inner = c1 * (x + c2 * x.pow(3.0))
+        val inner = c1 * (x + c2 * x.pow(3.0f))
         // Kotlin의 tanh는 Value 객체에 대해 정의되어 있지 않으므로, tanh를 직접 구현하거나 근사해야 합니다.
         // 여기서는 tanh(y) = (exp(2y) - 1) / (exp(2y) + 1) 공식을 사용합니다.
-        val exp2y = (inner * 2.0).exp()
-        val tanhInner = (exp2y - 1.0) / (exp2y + 1.0)
+        val exp2y = (inner * 2.0f).exp()
+        val tanhInner = (exp2y - 1.0f) / (exp2y + 1.0f)
 
-        return x * 0.5 * (1.0 + tanhInner)
+        return x * 0.5f * (1.0f + tanhInner)
     }
 
 
@@ -118,7 +118,7 @@ class Value(
         }
         buildTopo(this)
 
-        this.grad = 1.0
+        this.grad = 1.0f
         topo.reversed().forEach { it._backward() }
     }
